@@ -1,14 +1,14 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 29
-    buildToolsVersion "29.0.3"
+    compileSdkVersion 31
+//    buildToolsVersion "31.0.3"
 
 
     defaultConfig {
         applicationId "com.phone.aidl_client"
         minSdkVersion 21
-        targetSdkVersion 29
+        targetSdkVersion 31
         versionCode 1
         versionName "1.0"
 
@@ -23,7 +23,7 @@ android {
         }
     }
 
-    //开启Java1.8 能够使用lambda表达式
+    //开启Java11 能够使用lambda表达式
     compileOptions {
         sourceCompatibility JavaVersion.VERSION_1_8
         targetCompatibility JavaVersion.VERSION_1_8
