@@ -12,13 +12,13 @@ tasks.withType(JavaCompile) {
 
 
 android {
-    compileSdkVersion 28
-    buildToolsVersion "28.0.3"
+    compileSdkVersion 31
+//    buildToolsVersion "31.0.3"
 
 
     defaultConfig {
         minSdkVersion 21
-        targetSdkVersion 28
+        targetSdkVersion 31
         versionCode 1
         versionName "1.0"
 
@@ -61,7 +61,7 @@ android {
         experimental = true
     }
 
-    //开启Java1.8 能够使用lambda表达式
+    //开启Java11 能够使用lambda表达式
     compileOptions {
         sourceCompatibility JavaVersion.VERSION_1_8
         targetCompatibility JavaVersion.VERSION_1_8
@@ -129,9 +129,10 @@ dependencies {
 
     implementation files('libs/BASE64Decoder.jar')
     testApi 'junit:junit:4.12'
-    androidTestApi 'androidx.test:runner:1.3.0'
-    androidTestApi 'androidx.test.espresso:espresso-core:3.3.0'
-    api 'androidx.core:core-ktx:1.3.2'
+    androidTestApi 'androidx.test:runner:1.4.0'
+    androidTestApi 'androidx.test.espresso:espresso-core:3.4.0'
+    api 'androidx.core:core:1.6.0'
+    api 'androidx.core:core-ktx:1.6.0'
 
     //Mvvm模式框架导入
     def lifecycle_version = "2.2.0"
