@@ -4,7 +4,6 @@ import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
 import android.media.FaceDetector;
 import android.os.Handler;
-import android.view.View;
 import android.widget.TextView;
 
 import androidx.appcompat.widget.Toolbar;
@@ -34,19 +33,13 @@ public class ThreadPoolActivity extends BaseRxAppActivity {
     private TextView tevStartThreadPool2;
     private TextView tevStopThreadPool2;
 
-//    private ExecutorService excutor0;
     private ExecutorService excutor;
     private ExecutorService excutor2;
-    //    private FutureTask<String> futureTask;
     private Handler handler;
     private Vector<Bitmap> vector = new Vector<>();
 
-
-//    private Future<?> future0;
     private Future<?> future;
     private Future<?> future2;
-//    private MineAsyncTask mineAsyncTask;
-
 
     @Override
     protected int initLayoutId() {
@@ -55,47 +48,6 @@ public class ThreadPoolActivity extends BaseRxAppActivity {
 
     @Override
     protected void initData() {
-//        futureTask = new FutureTask<String>(new Callable<String>() {
-//            @Override
-//            public String call() throws Exception {
-//                LogManager.i(TAG, "startThreadPool*****");
-//                for (int i = 0; i < 5; i++) {
-//                    BitmapFactory.Options mOption = new BitmapFactory.Options();
-//                    mOption.inPreferredConfig = Bitmap.Config.RGB_565;
-//                    Bitmap mBitmap = null;
-//                    if (i == 0){
-//                        mBitmap = BitmapFactory.decodeResource(getResources(), R.mipmap.picture, mOption);
-//                    } else if(i == 1){
-//                        mBitmap = BitmapFactory.decodeResource(getResources(), R.mipmap.picture2, mOption);
-//                    } else if(i == 2){
-//                        mBitmap = BitmapFactory.decodeResource(getResources(), R.mipmap.picture3, mOption);
-//                    } else if(i == 3){
-//                        mBitmap = BitmapFactory.decodeResource(getResources(), R.mipmap.picture5, mOption);
-//                    } else {
-//                        mBitmap = BitmapFactory.decodeResource(getResources(), R.mipmap.picture8, mOption);
-//                    }
-//                    vector.add(mBitmap);
-//                }
-//                initFaceRecognition(vector);
-//
-//                Timer timer = new Timer();
-//                TimerTask timerTask = new TimerTask() {
-//                    @Override
-//                    public void run() {
-//                        BitmapFactory.Options mOption = new BitmapFactory.Options();
-//                        mOption.inPreferredConfig = Bitmap.Config.RGB_565;
-//                        Bitmap mBitmap = BitmapFactory.decodeResource(getResources(), R.mipmap.picture3, mOption);
-//                        vector.add(mBitmap);
-//                        LogManager.i(TAG, "startThreadPool vector*****" + vector.size());
-//                    }
-//                };
-//                timer.schedule(timerTask, 300);
-//                return "";
-//            }
-//        });
-
-
-//        mineAsyncTask = new MineAsyncTask();
     }
 
     @Override
@@ -108,68 +60,19 @@ public class ThreadPoolActivity extends BaseRxAppActivity {
         tevStopThreadPool2 = (TextView) findViewById(R.id.tev_stop_thread_pool2);
         setToolbar(false, R.color.color_FFE066FF);
 
-        tevStartThreadPool.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-//                excutor0 = Executors.newSingleThreadExecutor();
-//                future0 = excutor0.submit(new Runnable() {
-//                    @Override
-//                    public void run() {
-//                    }
-//                });
-                startThreadPool();
-            }
+        tevStartThreadPool.setOnClickListener(v -> {
+            startThreadPool();
         });
-        tevStopThreadPool.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                stopThreadPool();
-            }
+        tevStopThreadPool.setOnClickListener(v -> {
+            stopThreadPool();
         });
-        tevStartThreadPool2.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                startThreadPool2();
-            }
+        tevStartThreadPool2.setOnClickListener(v -> {
+            startThreadPool2();
         });
-        tevStopThreadPool2.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                stopThreadPool2();
-            }
+        tevStopThreadPool2.setOnClickListener(v -> {
+            stopThreadPool2();
         });
     }
-
-//    private class MineAsyncTask extends AsyncTask<String, Integer, String> {
-//
-//        @Override
-//        protected void onPreExecute() {
-//            super.onPreExecute();
-//
-//        }
-//
-//        @Override
-//        protected String doInBackground(String... strings) {
-//
-//            return "";
-//        }
-//
-//        @Override
-//        protected void onProgressUpdate(Integer... values) {
-//            super.onProgressUpdate(values);
-//        }
-//
-//        @Override
-//        protected void onPostExecute(String s) {
-//            super.onPostExecute(s);
-//        }
-//
-//        @Override
-//        protected void onCancelled() {
-//            super.onCancelled();
-//
-//        }
-//    }
 
     /**
      * 记一次Java参数传递时值发生变化问题（传Vector，就是线程安全的List）
@@ -187,7 +90,7 @@ public class ThreadPoolActivity extends BaseRxAppActivity {
             } else if (i == 1) {
                 mBitmap = BitmapFactory.decodeResource(getResources(), R.mipmap.picture6, mOption);
             } else if (i == 2) {
-                mBitmap = BitmapFactory.decodeResource(getResources(), R.mipmap.picture7, mOption);
+                mBitmap = BitmapFactory.decodeResource(getResources(), R.mipmap.picture8, mOption);
             } else if (i == 3) {
                 mBitmap = BitmapFactory.decodeResource(getResources(), R.mipmap.picture14, mOption);
             } else if (i == 4) {
@@ -219,19 +122,13 @@ public class ThreadPoolActivity extends BaseRxAppActivity {
             @Override
             public void run() {
                 LogManager.i(TAG, "startThreadPool excutor currentThread name*****" + Thread.currentThread().getName());
-//                //1、vector是原vector传来的时候，当原vector发生变化，他还是会变化的
-//                initFaceRecognition(vector);
-
-                //2、如果vector是由原vector.clone()传来的时候，当原vector发生变化，他是不会变化的
+                //1、如果vector是由原vector.clone()传来的时候，当原vector发生变化，他是不会变化的
                 Vector<Bitmap> list = (Vector<Bitmap>) vector.clone();
                 initFaceRecognition(list);
             }
         });
         LogManager.i(TAG, "startThreadPool&&&*****");
 
-//        mineAsyncTask.execute();
-
-//        futureTask.run();
     }
 
     /**
@@ -253,16 +150,6 @@ public class ThreadPoolActivity extends BaseRxAppActivity {
         }
     }
 
-//    private void stopThreadPool0() {
-//        if (excutor0 != null && future0 != null && !excutor0.isShutdown()) {
-//            future0.cancel(true);
-//            if (!excutor0.isShutdown()) {
-//                excutor0.shutdownNow();
-//                LogManager.i(TAG, "stopThreadPool0 shutdownNow*****");
-//            }
-//        }
-//    }
-
     /**
      * （1）修改线程池状态为STOP
      * （2）不再接收任务提交
@@ -280,17 +167,6 @@ public class ThreadPoolActivity extends BaseRxAppActivity {
                 LogManager.i(TAG, "stopThreadPool shutdownNow*****");
             }
         }
-
-//        if (mineAsyncTask != null && !mineAsyncTask.isCancelled()) {
-//            mineAsyncTask.cancel(true);
-//            LogManager.i(TAG, "stopThreadPool shutdownNow*****");
-//        }
-
-
-//        if (futureTask != null && !futureTask.isCancelled()) {
-//            futureTask.cancel(true);
-//            LogManager.i(TAG, "stopThreadPool cancel*****");
-//        }
     }
 
     /**
@@ -332,7 +208,6 @@ public class ThreadPoolActivity extends BaseRxAppActivity {
             handler.removeCallbacksAndMessages(null);
             handler = null;
         }
-//        stopThreadPool0();
         stopThreadPool();
         stopThreadPool2();
         super.onDestroy();
