@@ -4,6 +4,8 @@ import android.os.Environment;
 
 import com.phone.common_library.BaseApplication;
 import com.phone.common_library.callback.OnCommonSingleParamCallback;
+import com.trello.rxlifecycle3.android.ActivityEvent;
+import com.trello.rxlifecycle3.components.support.RxAppCompatActivity;
 
 import java.io.BufferedOutputStream;
 import java.io.File;
@@ -16,7 +18,6 @@ import io.reactivex.ObservableEmitter;
 import io.reactivex.ObservableOnSubscribe;
 import io.reactivex.android.schedulers.AndroidSchedulers;
 import io.reactivex.annotations.NonNull;
-import io.reactivex.disposables.Disposable;
 import io.reactivex.functions.Consumer;
 import io.reactivex.schedulers.Schedulers;
 
@@ -31,8 +32,6 @@ public class ReadAndWriteManager {
 
     private static final String TAG = ReadAndWriteManager.class.getSimpleName();
     private static ReadAndWriteManager manager;
-//    private List<Disposable> disposableList;
-    private Disposable disposable;
 
     /**
      * 私有构造器 无法外部创建
@@ -55,39 +54,21 @@ public class ReadAndWriteManager {
     /**
      * 将内容写入sd卡中
      *
-     * @param fileName 要写入的文件名
-     * @param content  待写入的内容
+     * @param rxAppCompatActivity
+     * @param fileName            要写入的文件名
+     * @param content             待写入的内容
      * @throws IOException
      */
-    public void writeExternal(String fileName,
+    public void writeExternal(RxAppCompatActivity rxAppCompatActivity,
+                              String fileName,
                               String content,
                               OnCommonSingleParamCallback<Boolean> onCommonSingleParamCallback) {
-//        //获取外部存储卡的可用状态
-//        String storageState = Environment.getExternalStorageState();
-//        //判断是否存在可用的的SD Card
-//        if (storageState.equals(Environment.MEDIA_MOUNTED)) {
-//            //路径： /storage/emulated/0/Android/data/com.yoryky.demo/cache/yoryky.txt
-////            filename = context.getExternalCacheDir().getAbsolutePath() + File.separator + filename;
-//            filename = Environment.getExternalStorageDirectory().getAbsolutePath()
-//                    + File.separator
-//                    + "Android"
-//                    + File.separator
-//                    + fileName;
-//
-//            LogManager.i(TAG,"writeExternal");
-//            FileOutputStream outputStream = new FileOutputStream(filename, false);
-//            BufferedOutputStream bufferedOutputStream = new BufferedOutputStream(outputStream);
-//            bufferedOutputStream.write(content.getBytes());
-//            bufferedOutputStream.close();
-//        }
-
         LogManager.i(TAG, "writeExternal");
-//        disposableList = new ArrayList<>();
-        disposable = Observable.create(new ObservableOnSubscribe<Boolean>() {
+        Observable.create(new ObservableOnSubscribe<Boolean>() {
             @Override
             public void subscribe(@NonNull ObservableEmitter<Boolean> e) throws Exception {
                 LogManager.i(TAG, "Observable thread is*****" + Thread.currentThread().getName());
-                String FILEPATH = Environment.getExternalStorageDirectory().getAbsolutePath()
+                String FILEPATH = rxAppCompatActivity.getExternalCacheDir()
                         + File.separator
                         + "Mine";
                 File dirs = new File(FILEPATH);
@@ -110,6 +91,7 @@ public class ReadAndWriteManager {
                 e.onComplete();
             }
         }).subscribeOn(Schedulers.io())
+                .compose(rxAppCompatActivity.<Boolean>bindUntilEvent(ActivityEvent.DESTROY))
                 .observeOn(AndroidSchedulers.mainThread())
                 .subscribe(new Consumer<Boolean>() {
                     @Override
@@ -125,7 +107,6 @@ public class ReadAndWriteManager {
                         onCommonSingleParamCallback.onError("写入失败");
                     }
                 });
-//        disposableList.add(disposable);
     }
 
     /**
@@ -155,27 +136,6 @@ public class ReadAndWriteManager {
             inputStream.close();
         }
         return stringBuilder.toString();
-    }
-
-//    private boolean isSubscribe() {
-//        if (disposableList != null && disposableList.size() > 0) {
-//            return true;
-//        }
-//        return false;
-//    }
-
-    public void unSubscribe() {
-//        if (isSubscribe()) {
-//            for (int i = 0; i < disposableList.size(); i++) {
-//                disposable = disposableList.get(i);
-                if (disposable != null && !disposable.isDisposed()) {
-                    disposable.dispose();
-                    disposable = null;
-                }
-//            }
-//            disposableList.clear();
-//            disposableList = null;
-//        }
     }
 
 }
